스탯 타입을 int에서 long으로 변환 (형변환시 오류 예외처리)

+++ UnitUI/UnitHealthUI.cs
@@ -46,7 +46,7 @@ public sealed class UnitHealthUI
         UnitUI.Billboard();
     }
 
-    public void SetHP(int curHP, int maxHP)
+    public void SetHP(long curHP, long maxHP)
     {
         _sldHP.value = (float)curHP / maxHP;
     }
